fix: address review feedback on Configure Step dialog
Split BaseImageSection/GpuSection/CacheSection out of StepConfigDialog.tsx
into their own files under dialogs/sections/, one component per file. Stop
closing the whole dialog when Base Image is reset to default. Fix Tab
label color contrast in dark mode.

Also fixes two bugs found while chasing the dark-mode contrast issue,
scoped to the same files:
- Input.tsx discarded any caller-supplied sx prop whenever the field
  had no validation error, since the error-state sx was assigned after
  spreading {...rest} in the same JSX element. This silently broke
  GpuSection's disabledFieldSx on the GPU Count field.
- GpuSection's disabledFieldSx hardcoded a dark-theme-only white text
  color instead of a --jp-* variable, and now also covers the label
  and border colors for the disabled GPU Count/GPU Type fields.
